requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.4
mysql-connector-python==8.0.12
parsel==1.5.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
queuelib==1.5.0
Scrapy==1.5.1
selenium==3.14.0
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
zope.interface==4.5.0
